nim version with term rewriting
in Nim you can define term rewriting templates as shown here. The standard term for Fibonacci is `fib(n-1) + fib(n-2)` but this can be rewritten to `2 * fib(n-2) + fib(n - 3)` which improves the execution speed to a high degree. The boundaries have to be adjusted of course to handle the case of n-3.
The code can be compiled with `nim c -d:release fib_rewrite.nim`